Map not resized from default to configured.
Prior to this commit, the map was first set to its
default size, and then resized to the configured
values. This could be confusing for the user, be-
cause the resizing is seen in the html, and also
led to some problems because the OpenLayers Map
parameters were applied with the default size.